requirements.txt
joblib==1.2.0
pandas==2.0.0
numpy==1.24.2
imblearn==0.0
scikit-learn==1.2.2
xgboost==1.7.5
lightgbm==3.3.5
catboost==1.1.1
PyYAML==6.0
requests==2.28.2
matplotlib==3.7.1
seaborn==0.12.2
nltk==3.6.5
scipy==1.7.1
textblob==0.17.1
wordcloud==1.8.1